rename policy generator file, update tests
